refactor: split envelope.rs into focused classification and test modules
Break 481-line envelope.rs (70 LOC logic + 410 LOC tests) into:
- envelope/mod.rs: classification logic
- envelope/tests.rs: all 20 test cases

All 65 Slack crate tests pass. No public API changes.
